Use OutOfBoundsError for SVM and inner classes
Also fixed type hints for float hyperparameters in tests.

+++ tests/safeds/ml/classical/classification/test_gradient_boosting.py
@@ -39,8 +39,8 @@ class TestLearningRate:
         assert fitted_model._wrapped_classifier is not None
         assert fitted_model._wrapped_classifier.learning_rate == 2
 
-    @pytest.mark.parametrize("learning_rate", [-1, 0], ids=["minus_one", "zero"])
-    def test_should_raise_if_less_than_or_equal_to_0(self, learning_rate: int) -> None:
+    @pytest.mark.parametrize("learning_rate", [-1.0, 0.0], ids=["minus_one", "zero"])
+    def test_should_raise_if_less_than_or_equal_to_0(self, learning_rate: float) -> None:
         with pytest.raises(
             OutOfBoundsError,
             match=rf"learning_rate \(={learning_rate}\) is not inside \(0, \u221e\)\.",
